Move library columns settings to library sheet
Closes #3969. Dialog-ception.
